+++ packages/config-webpack/bin/server
@@ -2,7 +2,7 @@
 
 const execa = require('execa');
 const isDocker = require('is-docker');
-const { PORT } = require('../helpers');
+const { PORT } = require('../lib/helpers');
 
 const port = process.env.PORT || PORT;
 const args = process.argv.slice(2);
