+++ packages/config-webpack/package.json
@@ -7,7 +7,7 @@
     "config",
     "webpack"
   ],
-  "main": "index.js",
+  "main": "lib/index.js",
   "repository": {
     "type": "git",
     "url": "git@github.com:airbnb/nimbus.git"
@@ -39,5 +39,10 @@
     "webpack-cli": "3.3.10",
     "webpack-dev-server": "^3.9.0",
     "webpack-graphql-loader": "^1.0.2"
+  },
+  "devDependencies": {
+    "@types/html-webpack-plugin": "^3.2.1",
+    "@types/terser-webpack-plugin": "^2.2.0",
+    "@types/webpack-bundle-analyzer": "^2.13.3"
   }
 }
